fixed write past make array size

+++ tests/array_generators.c
@@ -71,7 +71,7 @@ array_fill_fp64(double* a, size_t size, void (*rseed)(uint64_t),
 
 
 uint64_t
-array_fill_int16_t(int32_t* a, size_t size, void (*rseed)(uint64_t),
+array_fill_int16_t(int16_t* a, size_t size, void (*rseed)(uint64_t),
                    uint64_t (*rand_uint64_t)(void))
 {
     time_seed seed = {.time = omp_get_wtime()};
